Move TemplatePageDefinition type from reducer to actions file
